+++ 05_フロントエンド/landing-page/src/components/organism/Header/Header.tsx
@@ -7,30 +7,20 @@ import { Logo } from "@/components/atom/Logo"
 import { TextHoverable } from "@/components/atom/TextHoverable"
 import { LinksHorizontal } from "@/components/molecule/LinksHorizontal"
 
-export const Header = (): JSX.Element => {
-  // TODO: brand, linksを受け取る
-  const brand = "Brand"
-  const links = [
-    {
-      label: "Home",
-      href: "/",
-    },
-    {
-      label: "Blog",
-      href: "/blog",
-    },
-    {
-      label: "About us",
-      href: "/about",
-    },
-  ]
+type HeaderProps = {
+  brandName: string
+  links: { label: string; href: string }[]
+}
+
+export const Header = (props: HeaderProps): JSX.Element => {
+  const { brandName, links } = props
   return (
     <div className={styles["wrapper"]}>
       <Container>
         <Box spaceX="lg">
           <div className={styles["inner"]}>
             <div className={styles["brand"]}>
-              <Logo>{brand}</Logo>
+              <Logo>{brandName}</Logo>
             </div>
             <div className={styles["links"]}>
               <LinksHorizontal>
