+++ 05_フロントエンド/landing-page/src/components/organism/Header/index.ts
@@ -1,1 +1,3 @@
-export * from "./Header"
+import { HeaderContainer } from "./HeaderContainer"
+
+export const Header = HeaderContainer
